Move benchmark to relevant package. Added bench task
