+++ bench_test.go
@@ -2,15 +2,12 @@ package redis_test
 
 import (
 	"bytes"
-	"errors"
-	"net"
 	"testing"
 	"time"
 
 	redigo "github.com/garyburd/redigo/redis"
 
 	"gopkg.in/redis.v3"
-	"gopkg.in/redis.v3/internal/pool"
 )
 
 func benchmarkRedisClient(poolSize int) *redis.Client {
@@ -275,71 +272,4 @@ func BenchmarkZAdd(b *testing.B) {
 			}
 		}
 	})
-}
-
-func benchmarkPoolGetPut(b *testing.B, poolSize int) {
-	dial := func() (net.Conn, error) {
-		return &net.TCPConn{}, nil
-	}
-	pool := pool.NewConnPool(dial, poolSize, time.Second, 0)
-
-	b.ResetTimer()
-
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			conn, _, err := pool.Get()
-			if err != nil {
-				b.Fatalf("no error expected on pool.Get but received: %s", err.Error())
-			}
-			if err = pool.Put(conn); err != nil {
-				b.Fatalf("no error expected on pool.Put but received: %s", err.Error())
-			}
-		}
-	})
-}
-
-func BenchmarkPoolGetPut10Conns(b *testing.B) {
-	benchmarkPoolGetPut(b, 10)
-}
-
-func BenchmarkPoolGetPut100Conns(b *testing.B) {
-	benchmarkPoolGetPut(b, 100)
-}
-
-func BenchmarkPoolGetPut1000Conns(b *testing.B) {
-	benchmarkPoolGetPut(b, 1000)
-}
-
-func benchmarkPoolGetRemove(b *testing.B, poolSize int) {
-	dial := func() (net.Conn, error) {
-		return &net.TCPConn{}, nil
-	}
-	pool := pool.NewConnPool(dial, poolSize, time.Second, 0)
-	removeReason := errors.New("benchmark")
-
-	b.ResetTimer()
-
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			conn, _, err := pool.Get()
-			if err != nil {
-				b.Fatalf("no error expected on pool.Get but received: %s", err.Error())
-			}
-			if err = pool.Replace(conn, removeReason); err != nil {
-				b.Fatalf("no error expected on pool.Remove but received: %s", err.Error())
-			}
-		}
-	})
-}
-
-func BenchmarkPoolGetRemove10Conns(b *testing.B) {
-	benchmarkPoolGetRemove(b, 10)
-}
-
-func BenchmarkPoolGetRemove100Conns(b *testing.B) {
-	benchmarkPoolGetRemove(b, 100)
-}
-
-func BenchmarkPoolGetRemove1000Conns(b *testing.B) {
-	benchmarkPoolGetRemove(b, 1000)
 }
